requirements.txt
python-telegram-bot==20.7
yt-dlp
openai-whisper
moviepy
torch
